+++ src/lib.rs
@@ -77,6 +77,18 @@
 //!
 //! `ImageWidget` draws an image on the screen.  `png` format has been tested, but `jpg` and `gif`
 //! formats should also work, as they are part of the Piston library.
+//!
+//! `PushButtonWidget` draws a `BoxWidget`, overlaying text on the top of it, justified in any
+//! direction you desired (Left, Center, Right)  It produces an `on_clicked` callback when a
+//! click is detected (with the left mouse button) inside the bounds of the widget (ie. it
+//! overrides the mouse button click inside event.)
+//!
+//! `ToggleButtonWidget` operates the same as a `PushButtonWidget`, but offers the ability to toggle
+//! selected state.  `on_selected` is the callback that is triggered with the selected state
+//! returned.
+//!
+//! `ProgressWidget` draws a progress bar on the screen, with a `BoxWidget` as its base.  The
+//! color of the progress bar and its background can be changed at any time.
 
 /// Main module containing the run loop for the UI components, containers for windows and
 /// `Widget` trait objects, and so on.  Contains the core elements required to build
@@ -98,4 +110,7 @@ pub mod event;
 /// - Text Widget (for drawing text)
 /// - Timer Widget (for performing timer operations)
 /// - Image Widget (for drawing images)
+/// - Push Button Widget (for creating an interactive button that can be clicked)
+/// - Toggle Button Widget (for creating an on/off button that can be toggled)
+/// - Progress Widget (for displaying progress of an operation)
 pub mod widget;